Version 0.0.107
- SVG 1.2/Tiny
- partial SVG 1.2/Full (May 2004 draft):
  - conflicts with 1.2/Tiny always resolved in favour of 1.2/Tiny
  - extensions parsed but not processed
  - not complete, nor will it ever be
- SVG 2.0 (August 2018) with:
  - December 2018 Filter Effects
  - April 2021 Animations draft
- SVG 2.0 (April 2021 draft) (2.1 to be?) with:
  - October 2019 Filter Effects
  - April 2021 Animations draft

+++ recipe/heater/type/type_enum_vec.h
@@ -169,7 +169,7 @@ template < e_type E, typename ENUM, typename CATEGORY, CATEGORY INIT >
     ::std::bitset < max_combinable_enum > bs, gs, said;
     if (enum_vec < E, ENUM, CATEGORY, INIT > :: good ())
     {   for (auto val : enum_vec < E, ENUM, CATEGORY, INIT > :: value_)
-        {   if (val > max_combinable_enum)
+        {   if (val > static_cast < int > (max_combinable_enum))
             {   PRESUME (false, __FILE__, __LINE__);
                 nits.pick (nit_not_combine, es_catastrophic, ec_type, val.name (), " is too big; abandoning constrained enum check");
                 ok = false; break; }
